bugfix reslove conflict
