+++ apps/galaxy/package.json
@@ -27,7 +27,7 @@
     "next-transpile-modules": "9.0.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.36.1",
+    "react-hook-form": "^7.39.5",
     "yup": "^0.32.11",
     "yup-password": "^0.2.2",
     "yup-phone-lite": "^1.1.2"
