+++ package.json
@@ -1,0 +1,39 @@
+{
+  "name": "cc-multiverse",
+  "version": "0.0.0",
+  "private": true,
+  "workspaces": [
+    "apps/*",
+    "packages/*"
+  ],
+  "scripts": {
+    "prepare": "npx husky install",
+    "build": "turbo run build",
+    "build:static": "turbo run build:static",
+    "dev": "turbo run dev --parallel",
+    "test": "turbo run test",
+    "test:watch": "turbo run test:watch",
+    "test:coverage": "turbo run test:coverage",
+    "eslint": "turbo run eslint",
+    "type-check": "turbo run type-check",
+    "storybook": "turbo run storybook",
+    "format": "prettier --write \"**/*.{ts,tsx,md}\""
+  },
+  "devDependencies": {
+    "@cc/eslint-config": "file:packages/eslint-config",
+    "@cc/jest-config": "file:packages/jest-config",
+    "@commitlint/cli": "^17.1.2",
+    "@commitlint/config-conventional": "^17.1.0",
+    "husky": "^8.0.1",
+    "prettier": "latest",
+    "turbo": "latest",
+    "webpack": "^5.74.0"
+  },
+  "engines": {
+    "node": ">=14.0.0"
+  },
+  "packageManager": "npm@8.5.0",
+  "dependencies": {
+    "@tanstack/react-table": "^8.7.0"
+  }
+}
